fix(catalog): code-review followups before merge
Senior code review caught several issues; this commit folds them in
before opening the PR rather than chaining a separate fix commit.

Workflow (.github/workflows/validate-catalog.yml):
  - Add explicit permissions: { contents: read, issues: write }. Without
    this, the issue-creation step 403s on repos with the GitHub-default
    restricted GITHUB_TOKEN scope (the post-Feb-2023 default). Matches
    cron.yml::alert-on-failure's existing permissions block.
  - Switch tracking-issue lookup from state: 'open' to state: 'all',
    prefer open / fall back to most-recently-closed, reopen-and-comment
    on closed matches. Mirrors the cron.yml pattern that was hardened
    after the #70 -> #72 churn on 2026-05-04 and is documented in
    CLAUDE.md as a project convention. Also bumps actions/upload-artifact
    @v4 -> @v5 to match the rest of the repo.

Live test (tests/integration/test_catalog_live.py):
  - Add in-test retry: up to 2 retries with 5s backoff on
    TransientFetchError before failing. A single Ashby/Greenhouse hiccup
    at 07:00 UTC would otherwise open a tracking issue that has nothing
    to do with broken catalog data. Done in-test (vs. via
    pytest-rerunfailures) to avoid adding a dev dependency for one
    nightly test.

Pytest hooks (tests/conftest.py + tests/integration/conftest.py):
  - Move pytest_addoption / pytest_collection_modifyitems for
    --catalog-live to the top-level tests/conftest.py. Pytest officially
    recommends pytest_addoption only at rootdir-level conftest; in a
    subdir it works only as long as collection roots above it pull it
    in (so `cd tests/integration && pytest` would have failed before).
  - Drop the pytest_configure hook that registered the catalog_live
    marker via addinivalue_line and instead register it in
    pyproject.toml [tool.pytest.ini_options] markers, alongside other
    project markers. One source of truth, works with --strict-markers.

pyproject.toml:
  - Drop orphaned 'live_api' (already orphaned pre-this-PR) and
    'nightly' (used only by tests/integration/test_default_slugs_live.py
    which this PR deleted) markers; replace with 'catalog_live' (the
    only marker the deleted/added tests actually use).

Frontend (FollowedCompaniesSection.tsx):
  - Gate aria-controls and aria-activedescendant on `open`. WAI-ARIA
    requires aria-controls to reference an existing element; the
    listbox <div id={listboxId}> only renders when open=true, so the
    closed-state aria-controls would point at a non-existent id and
    fail axe-core's aria-valid-attr-value check.

API (app/api/companies.py):
  - Drop redundant .is_(True) on the boolean column comparison; the
    bare column reference produces identical SQL and reads cleaner.

Verified:
  - 312 passed, 23 catalog-live skipped on default backend run.
  - 23 passed live-validation under --catalog-live.
  - 231 passed across 45 frontend test files.
  - TypeScript clean.
