+++ .github/workflows/validate-catalog.yml
@@ -9,6 +9,9 @@ jobs:
   validate:
     runs-on: ubuntu-latest
     timeout-minutes: 10
+    permissions:
+      contents: read
+      issues: write
     steps:
       - uses: actions/checkout@v5
       - uses: astral-sh/setup-uv@v8.1.0
@@ -27,7 +30,7 @@ jobs:
           | tee catalog.log
       - name: Upload pytest results
         if: always()
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: catalog-validation
           path: |
@@ -35,35 +38,46 @@ jobs:
             catalog.log
       - name: Open or update tracking issue on failure
         if: failure()
-        uses: actions/github-script@v7
-        with:
-          script: |
-            const fs = require('fs')
-            const log = fs.readFileSync('catalog.log', 'utf-8').slice(-3000)
-            const title = 'catalog: live validation failed'
-            const body = `Nightly catalog validation failed. Tail of pytest output:\n\n\`\`\`\n${log}\n\`\`\`\n\nWorkflow run: ${context.serverUrl}/${context.repo.owner}/${context.repo.repo}/actions/runs/${context.runId}`
+        env:
+          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          REPO: ${{ github.repository }}
+          RUN_URL: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
+        run: |
+          set -euo pipefail
+          title="[catalog] live validation failing"
+          # Search BOTH open and closed so a `closes #N` on a fix PR doesn't
+          # orphan the tracker (matches cron.yml::alert-on-failure pattern,
+          # which is in turn motivated by the #70 -> #72 churn on 2026-05-04).
+          # Prefer OPEN; fall back to most-recently-closed; reopen if needed.
+          existing=$(gh issue list --repo "$REPO" --state all \
+            --search "$title in:title" \
+            --json number,state,closedAt --limit 20 \
+            --jq '([.[] | select(.state == "OPEN")] + ([.[] | select(.state == "CLOSED")] | sort_by(.closedAt) | reverse))[0] // empty' \
+            || echo "")
+          log_tail=$(tail -c 3000 catalog.log)
+          body="Nightly catalog validation failed.
 
-            const issues = await github.rest.issues.listForRepo({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              state: 'open',
-              labels: ['catalog-validation'],
-              per_page: 20,
-            })
-            const existing = issues.data.find(i => i.title === title)
-            if (existing) {
-              await github.rest.issues.createComment({
-                owner: context.repo.owner,
-                repo: context.repo.repo,
-                issue_number: existing.number,
-                body,
-              })
-            } else {
-              await github.rest.issues.create({
-                owner: context.repo.owner,
-                repo: context.repo.repo,
-                title,
-                body,
-                labels: ['catalog-validation'],
-              })
-            }
+          Workflow run: $RUN_URL
+
+          Tail of pytest output:
+          \`\`\`
+          $log_tail
+          \`\`\`
+
+          First step in triage: open the failing entry in companies.yaml,
+          curl the (provider, slug) against the upstream board manually,
+          and either replace the slug or drop the row in a follow-up PR."
+          if [ -n "$existing" ]; then
+            num=$(echo "$existing" | jq -r '.number')
+            state=$(echo "$existing" | jq -r '.state')
+            if [ "$state" = "CLOSED" ]; then
+              gh issue reopen "$num" --repo "$REPO" \
+                --comment "Reopened by validate-catalog: a new live-validation run failed after this issue was closed."
+            fi
+            gh issue comment "$num" --repo "$REPO" --body "$body"
+          else
+            gh issue create --repo "$REPO" --title "$title" \
+              --label "catalog-validation" \
+              --body "$body" || \
+            gh issue create --repo "$REPO" --title "$title" --body "$body"
+          fi
