+++ pyproject.toml
@@ -53,6 +53,5 @@ select = ["E", "F", "I", "UP"]
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 markers = [
-    "live_api: marks tests as making real external API calls (deselect with '-m not live_api')",
-    "nightly: live external API checks; run only on the nightly CI job",
+    "catalog_live: hits real public ATS boards; only run with --catalog-live.",
 ]
