+++ tests/integration/conftest.py
@@ -105,32 +105,3 @@ async def auth_headers(seeded_user):
         algorithm="HS256",
     )
     return {"Authorization": f"Bearer {token}"}
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--catalog-live",
-        action="store_true",
-        default=False,
-        help=(
-            "Run the catalog-live validation tests against real public ATS boards. "
-            "Used by the nightly validate-catalog GitHub Actions workflow; off by default "
-            "for local + PR CI runs."
-        ),
-    )
-
-
-def pytest_configure(config):
-    config.addinivalue_line(
-        "markers",
-        "catalog_live: hits real public ATS boards; only run with --catalog-live.",
-    )
-
-
-def pytest_collection_modifyitems(config, items):
-    if config.getoption("--catalog-live"):
-        return
-    skip_live = pytest.mark.skip(reason="needs --catalog-live to run")
-    for item in items:
-        if "catalog_live" in item.keywords:
-            item.add_marker(skip_live)
